% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{add_geneScores_to_df}
\alias{add_geneScores_to_df}
\title{CMAPResults data.frame with gene-level scores}
\usage{
add_geneScores_to_df(cmap.result)
}
\arguments{
\item{cmap.result}{CMAPResults object}
}
\value{
A data.frame with gene-level scores in column 'Scores'
}
\description{
This function converts a CMAPResults object into a data.frame and adds the gene.level scores as an additional column
}
\author{
Thomas Sandmann
}

