% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gCMAPWeb-package.R
\docType{data}
\name{cmap-data}
\alias{cmap-data}
\alias{cmap1}
\alias{cmap2}
\alias{cmap3}
\alias{cmap4}
\alias{cmap5}
\title{Example reference datasets}
\format{cmap1 to 4 are \code{\linkS4class{NChannelSet}} objects each with 1000 simulated z-scores for 10 samples. cmap5 is a \code{\linkS4class{CMAPCollection}} with membership information about 1000 genes and 10 gene sets.}
\description{
Five example reference datasets containing simulated data to demonstrate the
functionality of the gCMAPWeb package. These objects were generated with the
\code{\link{exampleCMAP}} function using different seeds.
}
\details{
cmap1, cmap2, cmap3 and cmap5 contain data for 1000 human Entrez identifiers.
cmap1, cmap2 and cmap3 are \code{\linkS4class{NChannelSet}} objects with one channel, the 'z' AssayDataElement.
cmap4 is an \code{\linkS4class{NChannelSet}} with data for 1000 mouse Entrez identifiers in the 'z' and 'log_fc' channels.
cmap5 is a code{\linkS4class{CMAPCollection}}.
}
\keyword{datasets}

