% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_overview_plot}
\alias{create_overview_plot}
\title{Overview plot}
\usage{
create_overview_plot(effect.sample, effect.population, file.name,
  reference.name, main = "Distribution of similarity scores",
  xlab = "Effect size", col.set = "black", col.up = "#1B9E77",
  col.down = "blue", strip.cutoffs = c(-3, 3), strip.bounds = c(-6, 6),
  strip.col = c("#1B9E77", "white", "blue"), url.base = NULL,
  up.label = "Correlated", down.label = "Anti-correlated")
}
\arguments{
\item{effect.sample}{numeric, the scores for all gene sets in the reference database deemed significantly similar}

\item{effect.population}{numeric, the scores for all gene sites in the reference database}

\item{file.name}{character, path and name of the output file}

\item{reference.name}{name of the reference dabatase}

\item{main}{main title of the plot}

\item{xlab}{x-axis label}

\item{col.set}{color used for the significant samples in the density plot}

\item{col.up}{color used to indicate significantly correlated / enriched gene sets in the rug plot}

\item{col.down}{color used to indicate significantly anti-correlated / depleted gene sets in the rug plot}

\item{strip.cutoffs}{numeric vector, indicating the area of the heat-strip set to the intermediat strip.col. Default=c(-3,3)}

\item{strip.bounds}{numeric vector, indicating the cutoffs above / below which the color of the heat-strip is set to maximum. Default=c(-6,6)}

\item{strip.col}{character vector with three elements, indicating high, neutral and low scores, respectively. Default=c("#1B9E77", "white", "blue")}

\item{url.base}{path the location of the output files relative to the gCMAPWeb home directory}

\item{up.label}{character, legend label for positive scores}

\item{down.label}{character, legend label for negative scores}
}
\value{
None. Pdf and png versions of the overview file are written to file.
}
\description{
This function creates an overview density plot summarizing the similarity scores obtained for all instances in the reference database.
}
\author{
Thomas Sandmann
}

