\name{create_profile_ExpressionSet}
\alias{create_profile_ExpressionSet}
\title{ExpressionSet creator}
\usage{
  create_profile_ExpressionSet(post, conf_data)
}
\arguments{
  \item{post}{list, POST component of the Rook request}

  \item{conf_data}{list, the configuration data as returned
  by the read_config_file function}
}
\value{
  ExpressionSet object
}
\description{
  ExpressionSet creator
}
\author{
  Thomas Sandmann
}

