\name{create_zip}
\alias{create_zip}
\title{Export gCMAPWeb report as zip archive}
\usage{
  create_zip(tmp_filename, out.dir)
}
\arguments{
  \item{tmp_filename}{character, name of the per-session
  result directory to be archived}

  \item{out.dir}{character, output directory to save zip
  archive into}
}
\value{
  Nothing, the zip archive is written to disk.
}
\description{
  This function exports the complete report in a zip file
}
\author{
  Thomas Sandmann
}

