% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{examplePost}
\alias{examplePost}
\title{Simulate POST request}
\usage{
examplePost(cmap, inputType = "non-directional", species = "human",
  array.platform = NULL, idType = "entrez", rows = 1000, add = 3,
  seed = 123)
}
\arguments{
\item{cmap}{NChannelSet object, e.g. generated by the exampleCMAP function}

\item{inputType}{character, one of directional, non-directional, profile}

\item{species}{character, should match a species supported by the loaded gCMAPWeb configuration file}

\item{array.platform}{character,  should match an array platform supported by the loaded gCMAPWeb configuration file}

\item{idType}{character, one of entrez, symbol or probe}

\item{rows}{integer, number of genes to include in a 'profile' query.}

\item{add}{numeric, score shift for scores in the first sample column. Scores for genes 1-50 will be shifted up, those for genes 51-100 down. Only used to simulate 'profile' queries.}

\item{seed}{numeric, seed for the random number generator. Default=123}
}
\value{
list, containg the same elements as a gCMAPWeb POST request
}
\description{
This function returns a simulated POST request, similar to those generated from user-input into the web interface.
}
\details{
This function is mainly intented for testing purposes.
}
\examples{
if( require( "org.Hs.eg.db", character.only=TRUE)){
  ## generate an example NChannelSet for human EntrezIds
  e <- exampleCMAP( universe="org.Hs.eg", rows=1000, cols=10)
  ## generate a matching example Post, simulating a non-directional query
  examplePost( cmap=e )
  ## or a directional query
  examplePost( cmap=e, inputType="directional")
}
}
\author{
Thomas Sandmann
}

