% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gCMAPWeb-package.R
\docType{package}
\name{gCMAPWeb-package}
\alias{gCMAPWeb-package}
\title{A web-interface for gene-set enrichment analysis}
\description{
A web-interface for gene-set enrichment analysis
}
\details{
This package provides a stand-alone web interface for gene-set enrichment analysis. Run either through R's internal Rhttpd server of deployed through an Apache2 webserver, gCMAPWeb allows users to search reference datasets stored in \code{\linkS4class{NChannelSet}} or \code{\linkS4class{CMAPCollection}} objects via three different methods:
\itemize{
  \item Directional queries performed by calculating the parametric JG score from the reference datasets
  \item Non-directional queries performed via Fisher's exact test
  \item Profile queires performed by calculating the parametric JG score from the query data
}
A web interace is populated based on a configuration file in YAML format and additional information from the reference dataset objects (if available).
For each query, a report is generated in html format, including graphs and tabular output. Extensive customization of the graphical user interface can easily be performed through global options.
}
\examples{
example( gCMAPWeb )
}
\author{
Thomas Sandmann
}
\references{
Jiang Z, Gentleman R., Extensions to gene set enrichment.
  Bioinformatics. 2007 Feb 1;23(3):306-13
  \url{http://www.ncbi.nlm.nih.gov/pubmed/17127676}
}
\keyword{package}

