% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_elements.R
\name{gene_density_chart_legend}
\alias{gene_density_chart_legend}
\title{Html code for the legend of pie the pie chart of the gene-level report}
\usage{
gene_density_chart_legend(text)
}
\arguments{
\item{text}{Character, text for the legend of the density plot on gene-level reports for non-directional and directional queries. Can be set as the global variabel "gene.density.legend".}
}
\value{
Character string with html code
}
\description{
Html code for the legend of pie the pie chart of the gene-level report
}
\author{
Thomas Sandmann
}

