% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_elements.R
\name{gene_density_profile_legend}
\alias{gene_density_profile_legend}
\title{Html code for the legend of the density charts of the gene-level reports from Profile queries}
\usage{
gene_density_profile_legend(text)
}
\arguments{
\item{text}{Character, text for the legend of the density plot on gene-level reports for profile queries. Can be set as the global variabel "gene.profile.legend".}
}
\value{
Character string with html code
}
\description{
Html code for the legend of the density charts of the gene-level reports from Profile queries
}
\author{
Thomas Sandmann
}

