% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_elements.R
\name{index_message}
\alias{index_message}
\title{Html code for text of the additional message on the index page (warning box)}
\usage{
index_message(text)
}
\arguments{
\item{text}{Character, text for the additional message on the index page (warning box)}
}
\value{
text Character string with html code
}
\description{
Html code for text of the additional message on the index page (warning box)
}
\author{
Thomas Sandmann
}

