% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createObject.R
\name{process_score_input}
\alias{process_score_input}
\title{Score parser}
\usage{
process_score_input(query_data, n.score.col = 1)
}
\arguments{
\item{query_data}{list of row-vectors, as generated by parse_file_input function}

\item{n.score.col}{integer, specifies the number of data columns (usually 1)}
}
\value{
matrix of expression scores
}
\description{
This function reformats score input uploaded or pasted by the user
}
\author{
Thomas Sandmann
}

