\name{signed_input_example}
\alias{signed_input_example}
\title{Html code for generating and example query button on the signed gene set submission page}
\usage{
  signed_input_example()
}
\value{
  Character string with html code
}
\description{
  Content can be provided / modified through the global
  parameters signed.input.example.popover ( text for
  popover help ) signed.input.example.up (up-regulated gene
  identifiers) signed.input.example.down (down-regulated
  gene identifiers)
}
\author{
  Thomas Sandmann
}

