% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_elements.R
\name{single_input_example}
\alias{single_input_example}
\title{Html code for generating and example query button on the single gene query submission page}
\usage{
single_input_example()
}
\value{
Character string with html code
}
\description{
Content can be provided / modified through the global parameters
single.gene.example.popover ( text for popover help )
single.gene.example (gene identifiers)
}
\author{
Thomas Sandmann
}

