% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_elements.R
\name{create_heatmap_legend}
\alias{create_heatmap_legend}
\title{Legend for overview heatmap}
\usage{
create_heatmap_legend(text)
}
\arguments{
\item{text}{Character, text for the legend of the density plot the main report page. Can be set as the global variabel "gene.set.legend".}
}
\value{
Character
}
\description{
Content for the legend of the overview heatmap plot displayed on the main results page for directional and non-directional queries
}
\author{
Thomas Sandmann
}

