% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_elements.R
\name{create_overview_legend}
\alias{create_overview_legend}
\title{Legend for overview density plot}
\usage{
create_overview_legend(text)
}
\arguments{
\item{text}{Character, text for the legend of the density plot the main report page. Can be set as the global variabel "gene.set.legend".}
}
\value{
Character
}
\description{
Content for the legend of the overview density plot displayed on every main results page
}
\author{
Thomas Sandmann
}

