% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createObject.R
\name{create_profile_ExpressionSet}
\alias{create_profile_ExpressionSet}
\title{ExpressionSet creator}
\usage{
create_profile_ExpressionSet(post, conf_data)
}
\arguments{
\item{post}{list, POST component of the Rook request}

\item{conf_data}{list, the configuration data as returned by the read_config_file function}
}
\value{
ExpressionSet object
}
\description{
This function creates an ExpressionSet object from user-specified identifiers and scores.
}
\author{
Thomas Sandmann
}

