\name{index_message}
\alias{index_message}
\title{Html code for text of the additional message on the index page (warning box)}
\usage{
  index_message(text = getOption("index.message", default = NULL))
}
\arguments{
  \item{text}{Character, text for the additional message on
  the index page (warning box)}
}
\value{
  text Character string with html code
}
\description{
  Html code for text of the additional message on the index
  page (warning box)
}
\author{
  Thomas Sandmann
}

