% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_elements.R
\name{index_quote}
\alias{index_quote}
\title{Html code for quote on the index page}
\usage{
index_quote(text)
}
\arguments{
\item{text, }{Character, the content of the blockquote field on the index.rhtml page. Can be set as the global variabel "index.quote".}
}
\value{
Character string with html code
}
\description{
Html code for quote on the index page
}
\author{
Thomas Sandmann
}

