% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_elements.R
\name{index_subtitle}
\alias{index_subtitle}
\title{Html code for subtitle of the hero unit of the index.rhtml page}
\usage{
index_subtitle(text)
}
\arguments{
\item{text}{Character, subtitle of the hero unit on the index.rhtml page. Can be set as the global variabel "index.sub".}
}
\value{
Character string with html code
}
\description{
Html code for subtitle of the hero unit of the index.rhtml page
}
\author{
Thomas Sandmann
}

