\name{index_text}
\alias{index_text}
\title{Html code for text of the hero unit of the index.rhtml page}
\usage{
  index_text(text = getOption("index.text", default = "Compare your favorite genes to a reference database of differential expression experiments"))
}
\arguments{
  \item{text}{Character, text of the hero unit on the
  index.rhtml page. Can be set as the global variabel
  "index.text".}
}
\value{
  Character string with html code
}
\description{
  Html code for text of the hero unit of the index.rhtml
  page
}
\author{
  Thomas Sandmann
}

