% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_elements.R
\name{index_title}
\alias{index_title}
\title{Html code for title of the hero unit of the index.rhtml page}
\usage{
index_title(text = getOption("index.main", default = "gConnectivity Map"))
}
\arguments{
\item{text}{Character, main title of the hero unit on the index.rhtml page. Can be set as the global variabel "index.main".}
}
\value{
Character string with html code
}
\description{
Html code for title of the hero unit of the index.rhtml page
}
\author{
Thomas Sandmann
}

