\name{profile_input_example}
\alias{profile_input_example}
\title{Html code for generating an example query button on the profile submission page}
\usage{
  profile_input_example()
}
\value{
  Character string with html code
}
\description{
  Content can be provided / modified through the global
  parameters profile.input.example.popover ( text for
  popover help ) profile.input.example (gene identifiers
  and scores)
}
\author{
  Thomas Sandmann
}

